change: Simplify graphics state and mesh uploading significantly
This greatly reduces the amount of API traffic when uploading chunks by
allowing all the layers to a chunk mesh to be allocated/uploaded in one
buffer block. It also gives the individual renderer backend more control
over how meshes are stored in memory, allowing specialized optimizations
to be used.

Additional work is needed to abstract the mesh generation and storage
concepts here, but this is better than what it was before.